The Event class has been rewritten. It is now an Entity for the events table in the database. Hibernate and Spring Data JPA have been connected. The EventDao class has been completely rewritten, and now all data is retrieved from the database using transactions.

+++ src/main/java/org/rockend/eco_calendar_web_application_demo/controller/CommonController.java
@@ -98,7 +98,7 @@ public class CommonController {
     }
 
     @PostMapping("/delete-event")
-    public String deleteEvent(@RequestParam("id") long eventId) {
+    public String deleteEvent(@RequestParam("id") int eventId) {
         eventService.deleteEvent(eventId);
 
         return "redirect:/admin-panel";
